examples/rpm/requirements.txt
pygame==1.9.2b6
pyserial==3.1.1

